refacotr: seperate useExpectationPage hook (#68)
